refactoring movie details screen and fix if not trailer
