+++ lib/ui/widgets/movie_details/components/movie_description_widget.dart
@@ -1,0 +1,50 @@
+import 'package:comics_db_app/app_colors.dart';
+import 'package:comics_db_app/ui/widgets/movie_details/movie_details_model.dart';
+import 'package:provider/provider.dart';
+import 'package:flutter/material.dart';
+
+class DescriptionWidget extends StatelessWidget {
+  const DescriptionWidget({
+    Key? key,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    final model = Provider.of<MovieDetailsModel>(context, listen: true);
+    return Padding(
+      padding: const EdgeInsets.only(left: 20.0, right: 20.0, top: 20.0),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          const Text(
+            'Описание',
+            style: TextStyle(
+              fontSize: 21,
+              fontWeight: FontWeight.w600,
+              // TODO rename text
+              color: AppColors.genresText,
+            ),
+          ),
+          const SizedBox(height: 8),
+          Row(
+            mainAxisAlignment: MainAxisAlignment.start,
+            children: [
+              Expanded(
+                // Добавить расстояние между строками
+                child: Text(
+                  model.movieDetails?.overview ?? 'Загрузка описания...',
+                  overflow: TextOverflow.ellipsis,
+                  maxLines: 4,
+                  style: const TextStyle(
+                    color: AppColors.genresText,
+                    fontSize: 12,
+                  ),
+                ),
+              ),
+            ],
+          ),
+        ],
+      ),
+    );
+  }
+}
