+++ lib/ui/widgets/movie_list/movie_list_widget.dart
@@ -120,8 +120,7 @@ class _MovieListWidgetState extends State<MovieListWidget> {
                     children: [
                       const Text('Популярные', style: TextStyle(color: AppColors.genresText, fontSize: 21, fontWeight: FontWeight.w600)),
                       InkWell(
-                        // onTap: () => Navigator.of(context).pushNamed(MainNavigationRouteNames.popularMovie),
-                        onTap: () {Navigator.of(context).pushNamed(MainNavigationRouteNames.popularMovie);},
+                        onTap: () => Navigator.of(context).pushNamed(MainNavigationRouteNames.popularMovie),
                           child: const Text('Все', style: TextStyle(color: AppColors.ratingText, fontSize: 15),)),
                     ],
                   ),
@@ -152,59 +151,6 @@ class _MovieListWidgetState extends State<MovieListWidget> {
             ),
                   ),
       );
-      //     // TODO FIX doesn't work
-      //      ListView.builder(
-      //           keyboardDismissBehavior: ScrollViewKeyboardDismissBehavior.onDrag,
-      //           padding: const EdgeInsets.only(top: 70.0),
-      //           itemCount: model.movies.length,
-      //           itemExtent: 165,
-      //           itemBuilder: (BuildContext context, int index) {
-      //             model.showedMovieAtIndex(index);
-      //             final movie = model.movies[index];
-      //             final posterPath = movie.posterPath;
-      //             return Padding(
-      //               padding: const EdgeInsets.symmetric(
-      //                   horizontal: 16.0, vertical: 10.0),
-      //               child: Stack(
-      //                 children: [
-      //                   Container(
-      //                     decoration: BoxDecoration(
-      //                         color: Colors.white,
-      //                         border:
-      //                         Border.all(color: Colors.black.withOpacity(0.2)),
-      //                         borderRadius:
-      //                         const BorderRadius.all(Radius.circular(20.0)),
-      //                         boxShadow: [
-      //                           BoxShadow(
-      //                             color: Colors.black.withOpacity(0.1),
-      //                             blurRadius: 8,
-      //                             offset: const Offset(0, 2),
-      //                           )
-      //                         ]),
-      //                     clipBehavior: Clip.hardEdge,
-      //                     child: Row(
-      //                       children: [
-      //                         posterPath != null ? Image.network(
-      //                             ApiClient.imageUrl(posterPath), width: 95)
-      //                             : const SizedBox.shrink(),
-      //                         const SizedBox(width: 15.0),
-      //                         MovieCard(movie: movie, model: model),
-      //                         const SizedBox(width: 10.0),
-      //                       ],
-      //                     ),
-      //                   ),
-      //                   Material(
-      //                     color: Colors.transparent,
-      //                     child: InkWell(
-      //                       borderRadius: BorderRadius.circular(20.0),
-      //                       onTap: () => model.onMovieTap(context, index),
-      //                     ),
-      //                   ),
-      //                 ],
-      //               ),
-      //             );
-      //           }),
-      //     ),
   }
 }
 
@@ -249,8 +195,11 @@ class _UpcomingMovieWidgetState extends State<_UpcomingMovieWidget> {
                     upcomingMovieModel.showedMovieAtIndex(index);
                     final upcomingMovie = upcomingMovieModel.movies[index];
                     final backdropPath = upcomingMovie.backdropPath;
-                    return backdropPath != null ? Image.network(ApiClient.imageUrl(backdropPath))
-                        : const SizedBox.shrink();
+                    return InkWell(
+                      onTap: () => upcomingMovieModel.onMovieTap(context, index),
+                      child: backdropPath != null ? Image.network(ApiClient.imageUrl(backdropPath))
+                          : const SizedBox.shrink(),
+                    );
                   }
               ),
             ),
